feat: infer service type from class
if you do not provide a type for a service, we can use the class's name to infer the type instead

closes #8
